fix: api-gateway : 클래스와 메서드의 이름이 같아 Spring이 Bean을 사용할때 발생하는 문제 해결
